Initial work on adding OpenAI classes.

+++ cqi_cpp/include/action.hpp
@@ -1,6 +1,6 @@
 class Action {
     public:
-        float value;
-        Action(float);
+        int value;
+        Action(int);
         ~Action();
 };
